projects/sd/src/python/requirements.txt
google-cloud-firestore

